% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{add_codes_for_field}
\alias{add_codes_for_field}
\title{Extract the codes for a given field and add them to the given dataframe}
\usage{
add_codes_for_field(results, field, colname)
}
\arguments{
\item{results}{The results list.}

\item{field}{The name of the field for which to extract codes.}

\item{colname}{The name of the new column which will contain the codes.}
}
\value{
A copy of the results with a code column added to each dataframe.
}
\description{
\code{add_codes_for_field} adds a column containing the codes for a given
field to the dataframes contained in the given results. The codes are
derived from the uris: specifically they are the last item in uri string
delimited with a colon. Where fields contain items for totals their uris do
not contain a corresponding uri for the total. This function handles that
case by creating a dummy code for the total (called "Total").
}
