% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{CACHE_API_KEY}
\alias{CACHE_API_KEY}
\title{API key cacge variable name}
\format{An object of class \code{character} of length 1.}
\usage{
CACHE_API_KEY
}
\description{
API key cacge variable name
}
\keyword{internal}
