% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_cube}
\alias{extract_cube}
\title{Extract the results of a query from the response for a given data cube}
\usage{
extract_cube(json, cube)
}
\arguments{
\item{json}{The results of the query as parsed json.}

\item{cube}{The number of the cube to extract.}
}
\value{
A list of the results for the given cube.
}
\description{
Extract the results of a query from the response for a given data cube
}
\keyword{internal}
