% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_items_df}
\alias{extract_items_df}
\title{Extract a dataframe of the item combinations represented in a query result}
\usage{
extract_items_df(items)
}
\arguments{
\item{items}{The list of items for a query result.}
}
\value{
A dataframe of the item combinations represented in the result.
}
\description{
Extract a dataframe of the item combinations represented in a query result
}
\keyword{internal}
