% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_results}
\alias{extract_results}
\title{Extract the results of a query from the response json}
\usage{
extract_results(json)
}
\arguments{
\item{json}{The results of the query as parsed json.}
}
\value{
A list of the results for the given cube.
}
\description{
\code{extract_results} processes the results of the query and extracts the
data in a format suitable for analysis.
}
