% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_table}
\alias{fetch_table}
\title{Send a table query and return the results}
\usage{
fetch_table(query, filename, simplify = TRUE)
}
\arguments{
\item{query}{Stat-Xplore query as a string.}

\item{filename}{The path to a text file containing a Stat-Xplore query.
This argument is not required but has priority: if a \code{filename} is
provided, the \code{query} argument is ignored.}

\item{simplify}{If TRUE and the result contains only one data cube, return
just that cube. The default value is TRUE.}
}
\value{
A list containing the results of the query, with one item per cube.
}
\description{
\code{fetch_table} sends a query to the table endpoint and returns the
response as a list of data cubes, each containing a list of data items for
each cube.
}
\details{
A query may be provided as a string or can be loaded from a file using the
\code{filename} argument. In most cases this is more convenient as
Stat-Xplore queries can be large and are most easily produced using the
table builder tool on the website.

Stat-Xplore returns one data cube of results for each measure included in
the query. In plain english, cubes that are produced from the same query
show different results for different measures but within the same table
structure.

If a query returns more than one cube, the function returns a list of cubes
where each cube is named according to the name of the measure it contains.
If there is only one cube in the results for a query and \code{simplify} is
set to TRUE (the default), a single cube is returned.

In the results of this function, each cube is represented as a list with the
following data items:

fields - the names of categorical variables represented in the data (list)
items - the names of the categories or levels within each field (list)
uris - the uris of the categories or levels within each field (list)
df - a dataframe of the categories and values in long form (tibble)
}
