% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the api key}
\usage{
get_api_key()
}
\value{
The api key as a string
}
\description{
\code{get_api_key} gets the api key that the package uses for communciation
with the server from the package cache.
}
