% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{has_api_key}
\alias{has_api_key}
\title{Check an api key has been set}
\usage{
has_api_key()
}
\value{
A boolean indicating whether the cache contains an api_key
}
\description{
\code{has_api_key} checks that an api key for communicating with the server
has been set in the package cache.
}
