% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{load_api_key}
\alias{load_api_key}
\title{Loads the api key from a file}
\usage{
load_api_key(filename)
}
\arguments{
\item{filename}{The path to a text file containing the api key.}
}
\description{
\code{load_api_key} is a convenience method that loads the api key from the
given file and sets it for the session.
}
