% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{request_error}
\alias{request_error}
\title{Report an error with an http request}
\usage{
request_error(response)
}
\arguments{
\item{response}{The text of the server reponse.}
}
\description{
Report an error with an http request
}
\keyword{internal}
