% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{request_table}
\alias{request_table}
\title{Send an http request with a query and return the response}
\usage{
request_table(query)
}
\arguments{
\item{query}{A valid Stat-Xplore query as a string.}
}
\value{
A list of the query results as parsed json.
}
\description{
\code{request_table} sends a query to the table endpoint, checks the
response, parses the response text as json, and returns the parsed data. If
the query syntax is not valid or the request fails for any other reason an
error is raised with the response text.
}
