% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the api key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{A valid Stat-Xplore api key.}
}
\description{
\code{set_api_key} sets the api key that the package uses for communiciation
with the server.
}
