% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statxplorer.R
\docType{package}
\name{statxplorer}
\alias{statxplorer}
\alias{statxplorer-package}
\title{statxplorer: Download data from the Stat-Xplore API}
\description{
The statxplorer package provides a suite of functions for downloading data
from the Department for Work and Pensions Stat-Xplore API.
}
