% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_table}
\alias{fetch_table}
\title{Send a table query and return the results}
\usage{
fetch_table(query, filename = NULL, simplify = TRUE)
}
\arguments{
\item{query}{Stat-Xplore query as a string.}

\item{filename}{The path to a text file containing a Stat-Xplore query.
This argument is not required but has priority: if a \code{filename} is
provided, the \code{query} argument is ignored.}

\item{simplify}{If TRUE and the results contain only one measure, don't nest
the measure and results data within unecessary lists.}
}
\value{
A list containing the results of the query, with one item per cube.
}
\description{
\code{fetch_table} sends a query to the table endpoint and returns the
response as a list containing the results. The results include a dataframe
of tidy data for each measure requested by the query.
}
\details{
A query can be provided as a string or can be loaded from a file using the
\code{filename} argument. This is sometimes convenient as Stat-Xplore
queries can be large and are most easily produced using the table builder
tool on the website, which exports queries as json text files.

Stat-Xplore returns one set of results for each measure included in a query.
Each set of results includes data for different measures on the same set of
observations.

The list of results has the following structure:

measure / measures - the names of the measures for each dataset (str / list)
fields - the names of categorical variables represented in the data (list)
items - the names of the categories or levels within each field (list)
uris - the uris of the categories or levels within each field (list)
df / dfs - a dataframe for each measure in long form (list / tibble)

If a query requests data on only one measure and \code{simplify} is
set to \code{TRUE}, then \code{relusts$measure} is a string containg the
measure name and \code{results$df} if a dataframe containing the results.

If a query requests data on more than one measure, or \code{simplify} is set
to \code{FALSE}, the measure names are stored as a list in
\code{results$measures} and the dataframes for each measure are stored as a
list in \code{results$dfs}.

\code{simplify} is \code{TRUE} by default.
}
