% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_table}
\alias{fetch_table}
\title{Send a table query and return the results}
\usage{
fetch_table(query, filename = NULL)
}
\arguments{
\item{query}{Stat-Xplore query as a string.}

\item{filename}{The path to a text file containing a Stat-Xplore query.
This argument is not required but has priority: if a \code{filename} is
provided, the \code{query} argument is ignored.}
}
\value{
A list containing the results of the query, with one item per cube.
}
\description{
\code{fetch_table} sends a query to the table endpoint and returns the
response as a list containing the results. The results include a dataframe
of tidy data for each measure requested by the query.
}
\details{
A query can be provided as a string or can be loaded from a file using the
\code{filename} argument. This is sometimes convenient as Stat-Xplore
queries can be large and are most easily produced using the table builder
tool on the website, which exports queries as json text files.

Stat-Xplore returns one set of results for each measure included in a query.
Each set of results includes data for different measures on the same set of
observations.

The list of results has the following structure:

measures - the names of the measures for each dataset (character)
fields - the names of categorical variables included in the data (character)
items - the names of the categories or levels within each field (list)
uris - the uris of the categories or levels within each field (list)
dfs - a dataframe for each measure with the data in long form (list)
}
